cpufreq: sysfs knobs for disabling hotplug during UI interaction
Provide the following sysfs knobs:
	/sys/devices/system/cpu/cpufreq/interactive
		core_lock_period, default 200000ns
		core_lock_count, default 0 (num_online_cpus)

Sysfs file core_lock_period controls the amount of time, in nanoseconds, to
hold some cores online.  It defaults to 200ms.  The number of cores to keep
online for that duration is controlled by core_lock_count, which defaults to 0.
When 0, we simply force num_online_cpus() to stay online; if greater than 0, we
force max(core_lock_count, num_online_cpus()) to stay online.

Note that you can write any value to core_lock_count, but it will get capped to
the number of cores actually available.

Change-Id: I6937e433c8343f8205e305d9da1a62f284b47f4f
Signed-off-by: Iliyan Malchev <malchev@google.com>